fiware-region-sanity-tests/requirements.txt
python-novaclient
python-neutronclient
nose
git+https://github.com/telefonicaid/nose-html-reporting.git
